Add check to File_Ogg for cut-off page headers
Should prevent logspam warning about unpack() failing on short input.

If a file was cut off in the middle of an Ogg page header, it might
try to parse the header and didn't have error checking inside. Now
checks that the header is the expected length, and reports back as
a parse failure which should throw an OggException which will either
get caught sanely or throw a backtrace that's more findable in logs.

Wikimedia bug: https://phabricator.wikimedia.org/T144483